python-app-template/{{cookiecutter.app_name}}/requirements.txt
PyYAML==3.13
prometheus_client==0.6.0
